# openapi.yaml
openapi: 3.1.1
info:
  title: Dream Diary API + MCP
  version: 1.2.0
  description: REST API and MCP endpoint for creating and managing dream entries with OAuth2 and legacy API key auth.
servers:
  - url: https://example.de
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: Legacy static bearer API key from `DREAM_DIARY_API_KEY`.
    oauth2:
      type: oauth2
      description: OAuth 2.0 Authorization Code with PKCE for Claude custom connectors.
      flows:
        authorizationCode:
          authorizationUrl: /oauth/authorize
          tokenUrl: /oauth/token
          scopes:
            dreamdiary.read: Read dream entries
            dreamdiary.write: Create, update and delete dream entries
  schemas:
    CreateDreamEntryRequest:
      type: object
      required:
        - text
      properties:
        text:
          type: string
          minLength: 1
          example: Ich bin durch einen hellen Wald geflogen.
        dreamDate:
          type: string
          format: date
          description: Optional dream date. If omitted, the API uses the current timestamp.
          example: 2026-03-23
    UpdateDreamEntryTextRequest:
      type: object
      required:
        - text
      properties:
        text:
          type: string
          minLength: 1
          example: Danach wurde der Himmel golden.
    DreamEntryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 12
        text:
          type: string
          example: Ich stand am Meer und sah zwei Monde.
        createdAt:
          type: string
          format: date-time
          example: 2026-03-23T07:45:30Z
        updatedAt:
          type: string
          format: date-time
          nullable: true
          example: 2026-03-23T08:01:05Z
    PagedDreamEntryResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/DreamEntryResponse'
        page:
          type: integer
          example: 0
        size:
          type: integer
          example: 20
        totalElements:
          type: integer
          format: int64
          example: 42
        totalPages:
          type: integer
          example: 3
        last:
          type: boolean
          example: false
...
    ApiError:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          example: 2026-03-23T10:15:30Z
        status:
          type: integer
          example: 400
        error:
          type: string
          example: Bad Request
        message:
          type: string
          example: Query parameter 'day' cannot be combined with 'start' or 'end'.
        path:
          type: string
          example: /api/dream-entries
        validationErrors:
          type: object
          additionalProperties:
            type: string
security:
  - bearerAuth: [ ]
  - oauth2: [ dreamdiary.read, dreamdiary.write ]
paths:
...
  /oauth/authorize:
    get:
      operationId: oauthAuthorizeScreen
      summary: OAuth authorization endpoint (Authorization Code + PKCE)
      parameters:
        - in: query
          name: response_type
          required: true
          schema:
            type: string
            enum: [ code ]
        - in: query
          name: client_id
          required: true
          schema:
            type: string
        - in: query
          name: redirect_uri
          required: true
          schema:
            type: string
            format: uri
        - in: query
          name: scope
          required: false
          schema:
            type: string
        - in: query
          name: state
          required: false
          schema:
            type: string
        - in: query
          name: code_challenge
          required: true
          schema:
            type: string
        - in: query
          name: code_challenge_method
          required: true
          schema:
            type: string
            enum: [ S256 ]
        - in: query
          name: resource
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Login form
        '400':
          description: Invalid request
    post:
      operationId: oauthAuthorizeSubmit
      summary: Submit OAuth login form and issue authorization code
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - response_type
                - client_id
                - redirect_uri
                - code_challenge
                - code_challenge_method
                - username
                - password
              properties:
                response_type:
                  type: string
                  enum: [ code ]
                client_id:
                  type: string
                redirect_uri:
                  type: string
                scope:
                  type: string
                state:
                  type: string
                code_challenge:
                  type: string
                code_challenge_method:
                  type: string
                  enum: [ S256 ]
                resource:
                  type: string
                username:
                  type: string
                password:
                  type: string
      responses:
        '302':
          description: Redirect to client redirect_uri with authorization code
        '200':
          description: Login form re-rendered (e.g. invalid credentials)
        '400':
          description: Invalid request
...
  /api/dream-entries:
    post:
      operationId: createDreamEntry
      summary: Create a dream entry
      security:
        - bearerAuth: [ ]
        - oauth2: [ dreamdiary.write ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDreamEntryRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DreamEntryResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    get:
      operationId: searchDreamEntries
      summary: Search dream entries
      description: If `day` is present, `start` and `end` are rejected with 400.
      security:
        - bearerAuth: [ ]
        - oauth2: [ dreamdiary.read ]
      parameters:
        - in: query
          name: day
          required: false
          schema:
            type: string
            format: date
          example: 2026-03-23
        - in: query
          name: start
          required: false
          schema:
            type: string
            format: date-time
          example: 2026-03-23T00:00:00Z
        - in: query
          name: end
          required: false
          schema:
            type: string
            format: date-time
          example: 2026-03-24T00:00:00Z
        - in: query
          name: q
          required: false
          schema:
            type: string
          example: meer
        - in: query
          name: page
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - in: query
          name: size
          required: false
          schema:
            type: integer
            minimum: 1
            default: 20
        - in: query
          name: sort
          required: false
          schema:
            type: string
            default: createdAt,desc
      responses:
        '200':
          description: Search result page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedDreamEntryResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/dream-entries/{id}:
    get:
      operationId: getDreamEntryById
      summary: Get one dream entry
      security:
        - bearerAuth: [ ]
        - oauth2: [ dreamdiary.read ]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Dream entry found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DreamEntryResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Dream entry not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    delete:
      operationId: deleteDreamEntry
      summary: Delete a dream entry
      security:
        - bearerAuth: [ ]
        - oauth2: [ dreamdiary.write ]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Deleted
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Dream entry not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/dream-entries/{id}/text:
    patch:
      operationId: updateDreamEntryText
      summary: Update only the text field of a dream entry
      security:
        - bearerAuth: [ ]
        - oauth2: [ dreamdiary.write ]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDreamEntryTextRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DreamEntryResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Dream entry not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'